refactor(serve): split the build out of ServeOptions into ServeBuildHost
API review follow-up. `ServeOptions` had grown to 99 members carrying two
different kinds of thing, and the member count was the least of it.

Ninety-one are flags — values the command parsed once. Seven were not options at
all: `autoInjectInitScriptArgs`, `gradleProjectRoot`, `gradleVariantArgs`,
`gradleBuildArgs`, `gradleProjects`, `runGradleTasks` and `discoverAndBuild` are
work the CLI does *on the server's behalf*. An option is answered at startup; a
build is invoked, can fail, and takes time. One interface said both.

Saying it separately makes the interesting question visible. A preview server
with no Gradle build behind it — serving only fetched bundles and catalogs, which
is already a supported mode — does not need a different `ServeOptions`. It needs
a different `ServeBuildHost`, and that is seven methods to write rather than
ninety-nine members to read past.

`autoInjectInitScriptArgs` moves with the build half despite reading like a
flag: it decides whether to inject the preview plugin into a project that does
not declare it, which is a build decision. It still binds `args` on the `:cli`
side, for the same reason as before.

`previewIdMatchesRequest` stays on `ServeOptions` — a selector rule belongs with
the selectors it reads.

Nothing about the Gradle-type argument changes: `ServeBuildHost` names `File`,
`String` and `PreviewModule` and no Gradle type, so the Tooling API stays on the
`:cli` side of the boundary. That reasoning moves to the new file, where it is
now the file's whole subject rather than a comment halfway down a 99-member
interface.

`serveInternalsUsedByCli` 12 -> 13: `ServeCommand` names one more serve type,
because it implements one more interface. The register grows by exactly the
seam this splits in two, which is the trade.

Tests: `:cli:serve:test` and `:cli:test` green under `--rerun-tasks` — forced,
because an ordinary run reported UP-TO-DATE after an earlier failure and an
ambiguous green is not a green. ktfmtCheck clean; seam gate OK at 13/16.
